MongoDB Desktop
Per [18f’s guide to naming
projects](https://pages.18f.gov/open-source-guide/naming-your-project/)
below.

We’ll move all of the scout-* modules soon and rename the repo.

> ### Before naming a project, think ‘Is this easy to understand?’
>
> To help users find and recognize your project, we recommend using
descriptive names that describe what your project does. You should
avoid acronyms and clever names/puns as those can make it hard to
search for projects.
>
> You should also do a quick search on the web for your project's name
to make sure that name isn't already being used by other software or
services, even if it's used in a different space, as it can be
confusing for new users. It’s also important to check with your
communications team before naming a project so that it can be cleared,
if need be.
>
> For example, if you were creating a template your coworkers could use
to create guides, a good name for the repo might be *guides-template*.
Bad names might include
*the-unnamed-project-that-makes-it-easy-to-build-stuff*, *temp-latte*,
or *guidestar*.
>
> Within the context of GitHub, the name of your project will always be
placed alongside its organizational owner's name. So it's unnecessary
to combine the two: instead of naming a project *my-organization-foo*,
you can simply name it *foo*.