INT-1104: Fix destroyed object exception on close.
When the window 'closed' event is fired, the window object is already
destroyed and requesting the id of the window would raise a 'object is
already destroyed error'. This moves the logic of removing the window
from the templates and removing the focus listener to the 'close' event
instead, and simply dereferencing the window on the 'closed' event.