Don't limit number of favorites.
If a user created 10 favorites, then no new connection would be able to
be used. This is because the connection collection was keeping the
entire connection list at a max size of 10, and a new connection would
get pruned when being created instead of being saved as a recent
connection.