:tada: :fire: Remove electron-compile
This module uses way too much "black magic" :dollar:
and has been a source of extreme confusion in determining
how things work or potential failure scenarios.

The only real functionality it's provided us with is
compiling less, which is replaced here via
[less' client-side rendering support](http://lesscss.org/#client-side-usage).
While suboptimal in the long-term, this provides us
with a short-term solution to remove electron-compile
and spend more time & energy on finding that long-term solution.

We can also speed up loading time by just checking in
index.html and removing the confusion of rending it from
a Jade template constantly.

@see http://npm.im/electron-compile