COMPASS-101: Fix electron build upgrade path:
Upgrades electron-packager to 8.0.0, which renames the executable on OSX
to the product name instead of "Electron". To support upgrades from
applications that were built before this upgrade, we add a symlink from
"Electron" to the new executable name.

Updates electron-prebuilt to electron.

Fixes the configuration to use the package.json from the current working
directory and not the package.json from the hadron-build module.