Compass 86 Atlas UX (#121)
* COMPASS-86: Revert changing SSL_DEFAULT to IFAVAILABLE

Downstream modules need time to decide to include such a change, i.e. they break hard with really cryptic errors.

This allows downstream modules to selectively enable this IFAVAILABLE option, we choose not to in Compass at this time, and can decide later if we drop this option entirely as we want to keep the UI simple.

* Update tests for changing SSL_DEFAULT `&ssl=prefer`

i.e. it should not be part of the default connection string at this time because it ties us to the system's root CA which isn't working for me so might be interesting to get working fully. Note: We still have some coverage of the `&ssl=prefer` option under the IFAVAILABLE tests.

* Refactor to lower cyclomatic complexity

* Add tests for ATLAS - mongodb.net

* COMPASS-86: Improve UX of connect to Atlas

We know Atlas is at *.mongodb.net, so use that information to pre-populate defaults that actually work out of the box, i.e. the user only has to also provide their password then it "just works" :)

* Add test for ATLAS 'works with a non-default secure password'

* Fix test so we support non-default passwords

Even more convenient 👍
Note: Abstracted to a separate _improveAtlasDefaults function to avoid the cyclomatic complexity warning.
See also https://github.com/mongodb-js/connection-model/pull/121#discussion_r83434447

* Add test 'does not false positive on hi.mongodb.net.my.domain.com'

https://github.com/mongodb-js/connection-model/pull/121#discussion_r83431822

* Add test for case insensitivity (RFC4343)

* Update isAtlas match to a case-insensitive regex at the domain's root