COMPASS-137: Reaonly view support in coll stats:
This allows the collection and index stats are in the top right of the
collection view to handle readonly views. They now display an "N/A" when
the user is viewing a readonly view.

This also removes the previous ampersand view for this component and
replaces it with a react component. The old collection stats ampersand
code has been removed.

The collection model no longer syncs on namespace change as this
component was the last part of the collection view that depended on the
amersand model.

This pull request has a dependency on:
https://github.com/10gen/compass/pull/578

That pull request MUST go in first.