More <Rule/> test warning fixes
    when category "range" is supplied
Warning: Failed prop type: The prop `serverVersion` is marked as required in `Rule`, but its value is `undefined`.
    in Rule
Warning: Failed prop type: The prop `isWritable` is marked as required in `Rule`, but its value is `undefined`.
    in Rule
Warning: Failed prop type: The prop `isWritable` is marked as required in `RuleCategoryRange`, but its value is `undefined`.
    in RuleCategoryRange