Merge test flags into test/README.md
After rereading Thomas’s comment more fully :)
Not sure if we need the current state of ‘npm test -- --packages’ not working to be documented here or in JIRA but happy to add back in. Also the two extra dashes are important and appear to have been omitted from the Slack message at this time.