Expand the --renderer test so it can handle integration tests
Tried several times to get this into the --unit tests, but our internal-packages tend to depend on “window” and “document” so that is not possible. Both those are genuinely available in the renderer process so it’s definitely a better home for these tests for now. We can decide later when we have more if they can or should be decoupled further.