COMPASS-701 First pass at Vega-Lite chart actions, store and tests (#766)
* First pass at chart actions, store and tests

* Add test to verify multiple actions get encoded in the channel state

A more naive implementation might perhaps have discarded keys after the first.

* :shirt: More readable test names

* Review: CHART_TYPE_ENUM: Use UPPER_SNAKE_CASE keys and kebab-case values

Brings back the lodash requirement but it’s required as JS design is simply not Python.
:yay: for tests :)

* Review: SPEC_TYPE_ENUM: Use UPPER_SNAKE_CASE keys and kebab-case values

* Review: Make _resetChart protected

* Review: Rename property -> channel

* Review: Rename selectField -> mapFieldToChannel action

* Review: Implement measurement as type

So we’ll have less trouble building a transformToVegaLiteSpec function later :)

* Review: AGGREGATE_FUNCTION_ENUM: Use UPPER_SNAKE_CASE keys and kebab-case values

* Review: MARK_PROPERTY_ENUM: Use UPPER_SNAKE_CASE keys and kebab-case values

* Rename to MARK_PROPERTY_CHANNEL_ENUM

* Review: MEASUREMENT_ENUM: Use UPPER_SNAKE_CASE keys and kebab-case values

* Review: Add the DETAIL_CHANNEL_ENUM

* Review: Generalise to CHART_CHANNEL_ENUM

Which is composed of the MARK_PROPERTY_CHANNEL_ENUM and the new DETAIL_CHANNEL_ENUM at this time.

* Review: bind onNamespaceChanged

* Freeze the CHART_CHANNEL_ENUM too

* Review: Add CHART_TYPE_CHANNELS and tests

Tests can catch obvious typos at least, like ‘require’ instead of ‘required’.

* Add getVegaLiteSpec, test and commented copy/paste manual test

* Update TODOs with JIRA ticket numbers

* Extract PackageManager setup into its own file

So I can possibly reuse it in tests where I’d like to test our listenToExternalStore wiring.

* Expand the --renderer test so it can handle integration tests

Tried several times to get this into the --unit tests, but our internal-packages tend to depend on “window” and “document” so that is not possible. Both those are genuinely available in the renderer process so it’s definitely a better home for these tests for now. We can decide later when we have more if they can or should be decoupled further.

* Finally fix tests and resolve COMPASS-631 conflicts

The unit/integration testing it in the renderer process without starting Compass was the challenging part :)

#retry (not yet JIRA-ticketed)
https://travis-ci.com/10gen/compass/jobs/63930156