:arrow_up: data-service@2.5.1
Includes :arrow_up: index-model@0.6.4 bug fix.

Note: As we don’t have a shrinkwrap.json or yarn.lock at this time committed on this releases branch, this is actually just a placeholder to retest that we didn’t break anything obvious on this tab as well as including this bug fix.