COMPASS-803: Backport COMPASS-586 COMPASS-788 icon changes (#792)
* COMPASS-586: Improve Linux appearance (logo in popups/about box, ...) (#762)

* COMPASS-586: Improve Linux appearance (logo in popups/about box, ...)

* assert.equals => assert.equal

* [Selectively cherry-picked from COMPASS 701 Charts]

* Extract PackageManager setup into its own file

So I can possibly reuse it in tests where I’d like to test our listenToExternalStore wiring.

* Expand the --renderer test so it can handle integration tests

Tried several times to get this into the --unit tests, but our internal-packages tend to depend on “window” and “document” so that is not possible. Both those are genuinely available in the renderer process so it’s definitely a better home for these tests for now. We can decide later when we have more if they can or should be decoupled further.

* COMPASS-788: mongodb:// in clipboard message has regressed on macOS / Red Hat (#787)

* Fix white screen and icon on macOS

If you have a string prefixed with `mongodb://`, e.g. every Atlas URL, present in your clipboard.

* Fix icon on Share Schema as JSON message box

Tested on a Red Hat VirtualBox VM, as per:
https://github.com/10gen/compass-internal-docs/pull/4

* Add regression test for the share-schema-as-json feature

Not unit testing the Ampersand View clipboard version as I have no idea how to do so.

* Update hadron-app => ampersand-app

As we don’t expect to back port b4a5a260aec9723ba09a87110467916958a31538

* Add ‘undefined’ to share-schema-as-json test

Not sure why this is ‘undefined’ on this branch but not master, but there’s a lot of hadron-app and other changes floating around so it probably doesn’t matter.