COMPASS-822: Remove hadron-app's dependency
Update hadron-app to use the dependency changes from [hadron-app
#1](https://github.com/mongodb-js/hadron-app/pull/1) that were
published in `hadron-app@0.0.6`. `lodash` has been removed as a
dependency in favor of using ES6 built-in language features.
`hadron-app` must be required early in the application's initialization
lifecycle, so removing external the external dependency avoids the
performance hit of resolving and loading `lodash`.