Inline migration requires
This works because we only save the preferences after the migrations have been run (presumably successfully, i.e. if they throw an error then we don't save the preferences change).

You can verify with a `debugger` here and a `debugger` inside src/app/index.js: Application.fetchPreferences(), or even a `throw new Error('Developer testing a migration error');` which sadly doesn't go to Bugsnag (would be nicer if it did, not sure if that's a bug).