Revert "Throw error for testing purposes"
Note: As discussed on Slack, no we won't get Bugsnag notifications for errors in migrations or user-preferences code, as that is before we have asked the user whether we're allowed to send such metrics. They'll have to contact us indirectly via other support channels.

"Decided in person that we wouldn't make time to fix it as fixes are non-trivial, so we'll just have to hope it doesn't happen and continue to be be "careful" with any and all migrations code"
https://mongodb.slack.com/archives/compass-scrum/p1489558248688923