Implement marky.mark for every dependency
Experiment to see if it gets us slightly better data on what we're spending our time doing.

It still doesn't handle the Compass depends on A and B, and both A and B depend on C so to remove the cost of C we'd need to remove it from both A and B problem, but does at least yield an interesting graph.