COMPASS-823 Migration speed (#861)
* Refactor app/migrations/index.js: var -> const

* Replace console.log calls from COMPASS 415 with marky

* npm install --save marky

Note: This new requirement is just 461 bytes (min+gz) and it's even recommended to "send them to your analytics provider".
For more details: https://github.com/nolanlawson/marky#why

Hence it's a dependency rather than a devDependency.

* Add Migrations marky timing

* Add semver check to skip migrations on the same Compass version

* Inline migration requires

This works because we only save the preferences after the migrations have been run (presumably successfully, i.e. if they throw an error then we don't save the preferences change).

You can verify with a `debugger` here and a `debugger` inside src/app/index.js: Application.fetchPreferences(), or even a `throw new Error('Developer testing a migration error');` which sadly doesn't go to Bugsnag (would be nicer if it did, not sure if that's a bug).

* Throw error for testing purposes

* Revert "Throw error for testing purposes"

Note: As discussed on Slack, no we won't get Bugsnag notifications for errors in migrations or user-preferences code, as that is before we have asked the user whether we're allowed to send such metrics. They'll have to contact us indirectly via other support channels.

"Decided in person that we wouldn't make time to fix it as fixes are non-trivial, so we'll just have to hope it doesn't happen and continue to be be "careful" with any and all migrations code"
https://mongodb.slack.com/archives/compass-scrum/p1489558248688923

* Instrument the main process migrations

Marky doesn't make sense as it's the main process, not the renderer process.

Compass main process migrations: 549.430ms

* Refactor main/migrations/index.js: var -> const

* Apply same inlining migrations fix as for the renderer process

Compass main process migrations: 424.379ms

So saves ~100 ms or so, worth taking as it's cheap.

#retry
'Compass Functional Tests for QueryBar #spectron when a MongoDB instance is running when using advanced query options when applying queries from the documents tab goes to the documents tab'
https://travis-ci.com/10gen/compass/jobs/68529519
https://travis-ci.com/10gen/compass/jobs/68555811
https://travis-ci.com/10gen/compass/jobs/68555803

'renders the home screen:
     Error: Promise was rejected with the following reason: timeout'
https://travis-ci.com/10gen/compass/jobs/68529938

* Replace `Rendering` with `Time to Connect rendered`

It's the more significant user statistic, the rest can be mined from the Chrome graph.

* Retrieve the lastKnownVersion without AmpersandModel

BEFORE
Compass main process migrations: 527.413ms

AFTER
Compass main process migrations: 151.401ms

* :shirt:

* Remove migration test speed code

We shouldn't need it anymore, easy enough to check out earlier commits too.

#retry
'Compass Functional Tests for QueryBar #spectron when a MongoDB instance is running when using advanced query options when applying queries from the documents tab goes to the documents tab'
https://travis-ci.com/10gen/compass/jobs/68692927
https://travis-ci.com/10gen/compass/jobs/68692924

'Compass Main Functional Test Suite #spectron when a MongoDB instance is running when viewing a collection when navigating to the indexes tab when creating an index not part of the schema fields adds a new field'
https://travis-ci.com/10gen/compass/jobs/68692928