COMPASS-912: Skip tests with #race and require 8 passes to merge to master (#905)
NOTE: This is a temporary change to unblock most other changes to master (such as the several other open PRs), and will be more completely resolved by COMPASS-857, e.g. via:

(1) revert at https://github.com/10gen/compass/pull/898 base, then
(2) conceptually cherry-pick the changes onto the latest version of that branch

* Decouple 'when the database name is valid' test

* Decouple 'when the collection name is valid' test

* Decouple 'does not close the insert documents modal on first press' test

* Make 'sorts the indexes by name' dependency explicit

It depends on the sometimes passes and sometimes fails 'allows another index to be added' test.

* Decouple 'when editing a document' test

* Skip tests identified to error at least twice with #race

We can review and reintroduce them on a case-by-case basis as time allows.
For details, please see:
https://github.com/10gen/compass/pull/897

* Skip some performance functional tests with #race

We have seen these fail in other contexts (e.g. COMPASS-806) as well as just failing for me on local again.
They should not slow the rest of the team down, ideally no team member should ever need to click Retry in Travis (however in practice, it should just be rare and due to circumstances genuinely beyond the team's control, such as AWS S3 being down).

* Backpedal on one skip

It's tightly coupled to the edit document flow.

* Add back MongoDB 3.0.x and 3.2.x

So as a team we have the choice to not just hit the magical Retry button which leads to this kind of mess, and catch such "sometimes" failures earlier in the development cycle.

* :shirt:

* Document and skip 'saves the changes to the document' #race

As detected in 3/40 runs by:
https://github.com/10gen/compass/pull/897#issuecomment-290321271

* Document the beforeEach hook of 'shares a "null" schema as JSON #race'

https://travis-ci.com/10gen/compass/jobs/70215764
https://travis-ci.com/10gen/compass/jobs/70396434

* Revert 'renders the network connections' test skip

* Forward port network connections change from COMPASS 857