COMPASS-990: Backport COMPASS 904 per plugin tooltip (#904)
NOTE: This excludes the new sidebar Create and Drop Database/Collection buttons as they were not present in our 1.6 release, but does cherry-pick the other secondary tooltips improvements.

* Remove unused this.Tooltip

It attempts to get an 'App.Tooltip' component from the AppRegistry that was not registered, and so ends up as undefined.

* Remove unused CREATE_DATABASE_ICON

It was decided a while back not to proceed with it as we already have the CREATE_DATABASE_BUTTON.

* Consolidate sidebar tooltips into top-level component

Anecdotally improves performance of the sidebar, might bring out the timeline when clicking the Sidebar Refresh button to be sure.

* abstract react-tooltip to HadronToolTip

* enable tooltip for databases-table

* enable toolltip for collections-table

* enable tooltip for crud

* enable tooltip for indexes

* enable tooltip for validation

* remove global tooltip from home

* add test for the collections table component

* add test for databases table component

* add test for indexes component

* add test for sampling message

* add test for rule builder component

* :shirt: