COMPASS-944 Part 3 - Swap and un-encode (#947)
* Revert 'prevent drag from encoding channel'

* Rename encodedChannel.field to fieldPath

* Provide encodedChannel.fieldName

Fixes a rendering issue of DraggableFields I just introduced when they are on the ChartPanel.
Now possible to see visually that DraggableFields from the ChartPanel are dropped onto the ChartPanel, just missing is the "swap" action :D

* Implement a basic "swap" behaviour

* Add a unit test for un-encoding a channel

* Implement the full swap behaviour

* Implement the 'x' to remove a field

As it depends on the implementation of `mapFieldToChannel(null, channel)` or some equivalently decoupled action like `unencodeChannel(channel)` to un-encode a channel, so I think it makes sense to bundle the review together.

* Avoid error if 'item.channelName' is undefined

i.e. No need to "swap" if item.channelName is not defined, as on first drag.

* Resolve regression in rendering a chart

https://vega.github.io/vega-lite/docs/encoding.html#field

* Add in &-can-drop from the tutorial

http://react-dnd.github.io/react-dnd/docs-tutorial.html

* Remove unused import from draggable-field enzyme test

* Make DraggableField.fieldName and .fieldPath required

* Update tests

- Use `fieldName` so as to not shadow the `name` builtin from somewhere else which is being set to the value `nodejs` for some reason =/
- Provide `field`, `fieldName` and/or `fieldPath` explicitly
- Test a nested fieldPath

* Fix enzyme test

I don't think PropTypes does conditional validation, e.g. there is yet another alternative but it illustrates how rare this semantic is:
https://www.npmjs.com/package/react-proptype-conditional-require

* Rename to chart.draggable-field.test.js

So it's easier to find when split into its own package later.

* :shirt:

* add swapEncodedChannels action and test