COMPASS-786: Use default distribution of 'compass':
After going through all the auto-updates and downloads code, it appears
we cannot support multiple products from the Github artifacts out of the
box as Nuts expects one product per repo. This commit reverts the
building and uploading of compass-enterprise, as it's not defined yet,
and defaults everything to plain compass for the existing auto-updates
and download center to work as is.