Return stub store when asking for nonexistent: (#12)
When the app registry is asked for a store that doesn't exist, its
previous behaviour was to return undefined. This could cause plugins
with issues to break the application because listen was not available to
be called by other plugins. Instead, we now return a fake store that can
be listened to so the application does not error. The app registry can
then be examined to see all names of stores that we asked for but were
missing and the number of times they were requested.