Document issues with depth/nestingLevel
e.g. how does ChartStore become aware the recursion or depth exists?
I'm looking at the `array_of_5_lng_lat_pairs` or `array_of_array_of_array_of_numbers` on the array_party data set.

It doesn't seem to be present when I inspect the FieldStore or the ChartStore.state.fieldsCache.
It appears to be present in a recursive structure in the SchemaStore after clicking Analyze:
-> SchemaStore.state.schema.fields[4].types