COMPASS-1319: Backport COMPASS-1318 Make EditableElement expand button hit box larger (#1119)
* :bug: COMPASS-1318 Make EditableElement expand button hit box larger

By using padding instead of margins, as margins are not clickable, but padding is.

* :bug: Add onClick handler to document keys

From review feedback:

> Quick question though: would it be easy to include the field name as a clickable area? I believe this got lost at some point. if it's a small change, can you just add it here?

Tested against both view mode and editing mode. AFAICT it seems to still respond intuitively, e.g. I can still double-click to select text independent of mode, single click to select a specific part of the key in edit mode, click and drag to select a part of the key in edit mode, etc.

More importantly, if I attempt to move the handler upwards into renderExpandable, then in edit mode, clicking on the editable key does NOT toggle the dropdown, which is an easy mistake I nearly made :)

* :fire: Remove unused renderStyle function

* :bug: Increase padding

The part between the arrow button and document key also should be clickable.