COMPASS-1288 reductions args (#1152)
* :tada: Add initial <ArrayReductionArg> component + tests

* :sparkles: Integrate <ArrayReductionArg> into <ArrayReductionPicker> and <DraggableField>

* :art: Style pass

* :tada: Handle validation

* :sparkles: Allow validators to coerce to different types

* :sparkles: Single-stage $arrayElemAt works

* :memo: Multiple dimensionality $arrayElemAt reductions probably not working

Most likely will end up in a follow up ticket as I'm not sure how to build an aggregation pipeline for these cases. The single dimensionality case working is worthy of a PR though, as for all I know at this time, this could even be a server bug.

* :memo: Improve REDUCTION_ARGS_TEMPLATE JSDoc

* :fire: Remove unused 'type' prop from <ArrayReductionArg>

* :rotating_light: Relax proptype warnings

Having them be numbers instead of strings when in the store makes building the agg pipelines easier.

* :bug: Handle Undo/Redo

* :art: Clean up argsTemplate in tests

It's a relic of the earlier implementation where it was a prop and trying to write these tests helped expose how silly it was :)

* :bug: Allow negative indexes

https://github.com/10gen/compass/pull/1152#issuecomment-315952183

* :art: More descriptive user-facing label

https://github.com/10gen/compass/pull/1152#discussion_r127877932

* :art: Improve readability with line breaks

https://github.com/10gen/compass/pull/1152#discussion_r127875753

* :memo: Add JSDoc for <ArrayReductionPicker> methods

* Defer 2D and 3D array tests until COMPASS-1413

* :rotating_light: Coerce to default value of ''

Warning: FormControl is changing an uncontrolled input of type text to be controlled. Input elements should not switch from uncontrolled to controlled (or vice versa). Decide between using a controlled or uncontrolled input element for the lifetime of the component. More info: https://fb.me/react-controlled-components

* :art: Resolve border style appearing embossed

I think it used the user-agent stylesheet, so just reuse the borders elsewhere on the DraggableField.

* :art: Add padding to make the user input more legible