COMPASS 1679 Reorder data-service setup to avoid instance refresh (#1192)
* Add some debugging statements

* Add DEBUG=* to Travis

* Drop travis to just the required functional subset

Should avoid the extremely verbose logs from enzyme tests as well, e.g.
https://api.travis-ci.com/jobs/85808677/log.txt?deansi=true&access_token=X8EJCyqCWec556MVGCK4qg

* Add more output

Very interesting Travis log from previous run:
https://travis-ci.com/10gen/compass/jobs/85809471

Parallels to COMPASS 1056 as well...

* :shirt:

* Move data service setup to before Compass starts

Avoids the need to fiddle with the .clickInstanceRefreshIcon() and .waitForInstanceRefresh() entirely, but if it does work, it would probably not resolve COMPASS 1056 as that would require a similar setup refactoring.

* Reverse debugging changes

* noop

* noop

* Revert "noop"