COMPASS 1740 Marky Tracking of CRUD (#5)
* npm i --save marky

Note: This might belong in Compass, just bringing it in in case I need it...

* Add marky to track 'Show N more fields' button

Adding a callback after setState has completed is very cool:
https://facebook.github.io/react/docs/react-component.html#setstate

* Add marky to track 'Expand All' button

* DocumentList: Add marky for load more, reset, remove, insert

* Typos and remove extraneous JSDoc params

* Add marky for Edit and Cancel

* Add marky for 'Handle update success'

* [Probably] Add marky for Clone

The npm linking has got too complicated for me at this point, React-Bootstrap is complaining and linking it and React appears to be insufficient. Might need React-Overlay as well, but going to stop chasing this and try publishing a release candidate instead and see if that works.

* Also npm link in react-bootstrap for Clone modal?

* Revert "Also npm link in react-bootstrap for Clone modal?"

This reverts commit 6304e7380b68a1de1a59a3fcace6eed4ba8ac6c2.

* 0.3.0-0