[Backport 1.8] COMPASS 2159 Switch to TravisCI container-based infrastructure (#1255)
* :fire: Remove #databases dropDatabase('music') call

It was not created by this test, so if it's needed it would be for earlier test runs? I don't see value in keeping it.

Plus - it might be how the chain of test timeouts gets broken? Let's find out with another test.

* :bug: Handle the case where app is undefined

This should be the second failure in at least one of the "main failure chain" pairs, i.e.
https://travis-ci.com/10gen/compass/jobs/92384423

(but not)
https://travis-ci.com/10gen/compass/jobs/92384429

* :bug: Handle app null as well as undefined

https://travis-ci.com/10gen/compass/jobs/92556220

* :construction: Try container-based infrastructure

Just seeing if it's a cheap win today.
https://blog.travis-ci.com/2014-12-17-faster-builds-with-container-based-infrastructure/

* :art: Remove clickDatabaseInSidebar/waitForSidebar calls

They are redundant as they are already done by goToCollection, which is at least a code smell worthwhile removing.

* noop

* Revert "noop"

* noop

* Revert "noop"

* noop

* Revert "noop"

* noop

* Revert "noop"

* noop

* Revert "noop"

#retry
  1) #collections when creating & deleting collections when viewing the database when creating a collection when deleting a collection pressing enter on correct collection name removes collection:
     Error: element ("[data-test-id='drop-collection-modal']") still not visible after 8000ms
https://travis-ci.com/10gen/compass/jobs/93287440

  1) #schema when applying a filter in the schema tab "before all" hook:
https://travis-ci.com/10gen/compass/jobs/93288256