COMPASS-2350: Fix edge case editing.
When going to the documents tab, navigating to the table view and back
to the list view, and then subsequently editing a document, some edit
functionality was broken.

This was because the component was still instantiated, received new
props, but didn't reset the listeners to the hadron-document events on
the new document.

Hard to debug, 1 line fix. We just needed to reset the listeners.