COMPASS-2563: Dont send metrics if instance failed
In the case of where an instance load failed, Compass would still try to
send metrics on the deployment without checking if the information was
there. This would cause the metrics send to fail with the vague
'Cannot read property length of null' error.

This fixes that, and allows the error to bubble up and be displayed
nicely to the user.