Merge pull request #3 from mongodb-js/COMPASS-3115_fix-windows-commandline-too-long
COMPASS-3115: Windows: AHHHHHHHHHHHHH. Fixed.

## The Fix

Do not pass `env: process.env` to `spectron.Application` constructor!

## Forensics

Found this buried deep in the chromedriver logs which can be enabled by passing the `chromeDriverLogPath` option to `spectron.Application` constructor.

```
[1540550592.876][DEBUG]: DevTools request: http://localhost:12853/json/version
The command line is too long.
[1540550593.908][DEBUG]: DevTools request failed
```

The devtools request is optional. The real problem is `The command line is too long.`! This is being caused because all `npm_*` env variables were being passed to spectron which converts these all to CLI args passed to chromedriver.  We shouldn't need to pass these explicitly as the subprocesses will auto inherit from the parent proc. Will need to do a bit more testing, but this mystery is finally looking solved!