COMPASS-3546: Fix Connections Requiring Password:
The connection model version was not updated in the data-service to the
models that both Compass and the data-service were using were actually
different. Compass was using an extended model and the data-service was
using the old undecorated model. This meant where the actuall connection
was happening in the data-service the model was not properly retrieving
the username and password from the secure storage.

This fixes that issue, and also removes the cyclic dependency between
the extended connection model and the data-service. The #test method was
not used in Compass thus the data-service connection testing
functionality in the model itself was no longer needed. This must only
happen in the data service.

This also brings the keytar versions in the data-service and Compass in
line. Could potentially fix COMPASS-3235. I don't see this behaviour
anymore but will give it some time. This also fixes the tests in the
data-service since the electron require in the extended models now
protect against failure to load electron in pure node envs.