Fix document list refresh on configureStore:
The previous fix for the tests didn't actually work as bind() returns a
function, and was unnecessary anyways since this is already set properly
for the refreshDocuments() call. The problem was that the namespace was
not set in the store so the insert calls were failing. This also adds an
addition option for testing called noRefreshOnConfigure which allows
tests to skip this stage and not have to deal with the store triggering
twice.