fix(import): COMPASS-4164: Preview table values always strings (#1914)
https://github.com/mongodb-js/compass-import-export/pull/120

## Description
(wasn't handling `Date` properly in example @mmarcon had on COMPASS-4164)

From mongodb-js/compass-import-export#120

> - chore: Upgrade to javascript-stringify@2 b5bdc7b Fixes #56
> - fix(import): Preview table values always strings 29148cd Fixes COMPASS-4164
> - fix(import): Hard limit reading fs for preview to first 20kb so preview always shows as fast as possible.
> 
> ![Screenshot 2020-03-09 12 18 44](https://user-images.githubusercontent.com/23074/76234690-6f985e00-6200-11ea-86a7-e902b2d0510f.png)

As a reminder, because we're now using a streaming backend, the total number of documents is likely to be off, especially for very large imports. So, 100% progress but still running is expected behavior. Devtools will log not only the input options that user chose but the estimator details as well:
![Screenshot 2020-03-09 12 40 32](https://user-images.githubusercontent.com/23074/76236618-580ea480-6203-11ea-89b9-1f91af4256a3.png)

Summary stats are also logged when the import completes successfully, by user selection or error:
![Screenshot 2020-03-09 12 41 20](https://user-images.githubusercontent.com/23074/76236622-593fd180-6203-11ea-9d29-a1eded665b11.png)

![Screenshot 2020-03-09 12 45 25](https://user-images.githubusercontent.com/23074/76236992-e4b96280-6203-11ea-8e4a-af6faf5214e0.png)