COMPASS-4478 - Support for updates in sharded collections (#306)
Include the shard key in the update query even if it has not
been modified, because the server requires us to do so.

This is enough to make updates work that do *not* touch the
shard key itself. For those, we’ll need to wait for the next
Node.js driver release.