fix: export all fields in CSV COMPASS-4332 (#235)
fast-csv autodetects columns based on the first row of data
when we don’t pass in the headers explicitly.

Start passing in the headers explicitly so that all selected
columns are included in the resulting CSV.

For achieving this goal, track the full list of available fields and
the selection/projection separately. When exporting, generate the
list of headers from the list of all available fields that match
the projection (e.g. `a.b.c` if `a.b` is selected in the UI).