fix(connection-model): make mongodb a proper dependency
This package depends on the mongodb driver not just as a dev
requirement, but also for its actual code.

I’ve spent some time trying to make it compatible so that it
would work with the 4.0 driver as well, but that seems like a
larger task after all and it probably makes more sense to just
do proper npm dependency declaration here, and make sure that
this package always gets the 3.x driver.