fix(ci): Switch compilation scripts in packages from prepare to prepublishOnly (#2356)
* fix(ci): Switch compilation scripts in packages from prepare to prepublishOnly

npm >= 7.20 started to run install lifecycle scripts automatically in workspace
subpackages and it is doing it without taking the topology of the packages in
account. This causes install/ci to fail because subpackages are trying to bootstrap
themselves out of order and fail because their workspace dependencies are not ready
yet. This patch switches all prepare scripts to prepublishOnly so that npm does not
run them on install, but they are still activated when running publish with lerna.

* chore: Split bootstrap and prepublishOnly into two separate commands

* chore: Change prepublishOnly to bootstrap in CI configs