fix(explain-plan): migrate to TS/away from Ampersand, fix sharding warning bug (#2494)
* chore(explain-plan): drop serverVersion from state

This does not have any actual impact on the UI.

* feat(explain-plan): replace explain-plan ampersand model

Replace the explain-plan-model Ampersand model and replace
it with a typescript helper package.

This new package matches the model broadly, except in a
few aspects:

- It automatically includes the SBE compatibiltiy helper,
  removing the need to apply it explicitly in compass-explain-plan.
- The “raw” original data property is exposed directly, not
  attached via monkey-patching in compass-explain-plan.
- `model.serialize()` is gone, the compass-explain-plan
  package now manually unpacks the properties it uses.
  This may not look very pretty, but at least it doesn’t
  hide the fact that we are using the object as an untyped
  bag of properties anymore.
- The child stage iterator helper is exposed, allowing some
  code redundancy between the packages to be removed.
- Support for pre-3.0 server explain plans is dropped.

* chore(explain-plan): replace usedIndex field to make it more meaningful

Replace the `usedIndex` field, which had an odd typing to begin
with, with a `usedIndexes` field that reflects all indices used
and what shards they were used on in a consistent way that matches
its actual usage.

* fix(explain-plan): do not display shard warning when inappropriate COMPASS-4770

This warning is supposed to tell people about situations in which
different shards used difference indices in their explain plan,
which is fine to warn about.

However, the check was fairly naïve and only warned whether multiple
indices were used at all. That can lead to false positives, since
using multiple indices by itself is not problematic.

Address this by grouping the used indices by the shards they were
run on, and seeing if these structures actually differ between
the individual shards.