fix(compass-schema): unambiguously display lat/long on map COMPASS-5526
GeoJSON (and thus MongoDB) stores coordinates as [Longitude, Latitude],
while elsewhere the reverse format [Latitude, Longitude] is common.

Compass displays the points on the map correctly, but reverses
the two values in the mouseover text.

Instead of switching to MongoDB/GeoJSON order, just clearly mark which
value is latitude and which is longitude.